+++ packages/nimble-blazor/NimbleBlazor/Components/NimbleThemeProvider.razor.cs
@@ -1,14 +1,25 @@
 ﻿using Microsoft.AspNetCore.Components;
+using Microsoft.JSInterop;
 
 namespace NimbleBlazor;
 
 public partial class NimbleThemeProvider : ComponentBase
 {
+    private ElementReference _themeProvider;
+    internal static string CheckThemeProviderValidityMethodName = "NimbleBlazor.ThemeProvider.checkValidity";
+    internal static string GetThemeProviderValidityMethodName = "NimbleBlazor.ThemeProvider.getValidity";
+
+    [Inject]
+    private IJSRuntime? JSRuntime { get; set; }
+
     [Parameter]
     public Direction? Direction { get; set; }
 
     [Parameter]
     public Theme? Theme { get; set; }
+
+    [Parameter]
+    public string? Lang { get; set; }
 
     [Parameter]
     public RenderFragment? ChildContent { get; set; }
@@ -18,4 +29,20 @@ public partial class NimbleThemeProvider : ComponentBase
     /// </summary>
     [Parameter(CaptureUnmatchedValues = true)]
     public IReadOnlyDictionary<string, object>? AdditionalAttributes { get; set; }
+
+    /// <summary>
+    /// Returns whether or not the theme provider configuration is valid.
+    /// </summary>
+    public async Task<bool> CheckValidityAsync()
+    {
+        return await JSRuntime!.InvokeAsync<bool>(CheckThemeProviderValidityMethodName, _themeProvider);
+    }
+
+    /// <summary>
+    /// Returns the validity state of the theme provider.
+    /// </summary>
+    public async Task<IThemeProviderValidity> GetValidityAsync()
+    {
+        return await JSRuntime!.InvokeAsync<ThemeProviderValidity>(GetThemeProviderValidityMethodName, _themeProvider);
+    }
 }