+++ packages/nimble-components/src/all-components.ts
@@ -44,6 +44,7 @@ import './table';
 import './table-column/anchor';
 import './table-column/date-text';
 import './table-column/enum-text';
+import './table-column/number-text';
 import './table-column/icon';
 import './table-column/text';
 import './tabs';