+++ packages/nimble-components/src/number-field/tests/number-field-matrix.stories.ts
@@ -39,8 +39,15 @@ const appearanceStates = Object.entries(NumberFieldAppearance).map(
 );
 type AppearanceState = (typeof appearanceStates)[number];
 
+const hideStepStates = [
+    ['', false],
+    ['Hide Step', true]
+] as const;
+type HideStepState = (typeof hideStepStates)[number];
+
 const component = (
     [disabledName, disabled]: DisabledState,
+    [hideStepName, hideStep]: HideStepState,
     [valueName, valueValue, placeholderValue]: ValueState,
     [errorName, errorVisible, errorText]: ErrorState,
     [appearanceName, appearance]: AppearanceState
@@ -51,18 +58,20 @@ const component = (
         value="${() => valueValue}"
         placeholder="${() => placeholderValue}"
         appearance="${() => appearance}"
+        ?hide-step="${() => hideStep}"
         ?disabled="${() => disabled}"
         error-text="${() => errorText}"
         ?error-visible="${() => errorVisible}"
     >
         ${() => errorName} ${() => appearanceName} ${() => valueName}
-        ${() => disabledName}
+        ${() => hideStepName} ${() => disabledName}
     </${numberFieldTag}>
 `;
 
 export const numberFieldThemeMatrix: StoryFn = createMatrixThemeStory(
     createMatrix(component, [
         disabledStates,
+        hideStepStates,
         valueStates,
         errorStates,
         appearanceStates