+++ packages/nimble-components/src/rich-text/editor/index.ts
@@ -26,6 +26,7 @@ import Paragraph from '@tiptap/extension-paragraph';
 import Placeholder from '@tiptap/extension-placeholder';
 import type { PlaceholderOptions } from '@tiptap/extension-placeholder';
 import Text from '@tiptap/extension-text';
+import HardBreak from '@tiptap/extension-hard-break';
 import { template } from './template';
 import { styles } from './styles';
 import type { ToggleButton } from '../../toggle-button';
@@ -356,6 +357,7 @@ export class RichTextEditor extends FoundationElement implements ErrorPattern {
                     placeholder: '',
                     showOnlyWhenEditable: false
                 }),
+                HardBreak,
                 customLink.configure({
                     // HTMLAttribute cannot be in camelCase as we want to match it with the name in Tiptap
                     // eslint-disable-next-line @typescript-eslint/naming-convention